requirements.txt
APScheduler==3.0.1
Flask==0.10.1
Flask-Login==0.2.11
Flask-Script==2.0.5
Flask-WTF==0.10.3
Jinja2==2.7.3
MarkupSafe==0.23
WTForms==2.0.1
Werkzeug==0.9.6
beautifulsoup4==4.3.2
futures==2.2.0
gspread==0.2.2
itsdangerous==0.24
peewee==2.4.5
python-dateutil==2.4.0
pytz==2014.10
requests==2.5.1
six==1.9.0
tzlocal==1.1.2
waitress==0.8.9
wsgiref==0.1.2
